+++ user-operator/src/test/java/io/strimzi/operator/user/UserOperatorConfigTest.java
@@ -4,6 +4,7 @@
  */
 package io.strimzi.operator.user;
 
+import io.strimzi.api.kafka.model.CertificateAuthority;
 import io.strimzi.operator.common.InvalidConfigurationException;
 import io.strimzi.operator.common.model.Labels;
 import org.junit.jupiter.api.Test;
@@ -28,8 +29,10 @@ public class UserOperatorConfigTest {
         envVars.put(UserOperatorConfig.STRIMZI_CA_NAMESPACE, "differentnamespace");
         envVars.put(UserOperatorConfig.STRIMZI_ZOOKEEPER_CONNECT, "somehost:2181");
         envVars.put(UserOperatorConfig.STRIMZI_ZOOKEEPER_SESSION_TIMEOUT_MS, "6000");
+        envVars.put(UserOperatorConfig.STRIMZI_CLIENTS_CA_VALIDITY, "1000");
+        envVars.put(UserOperatorConfig.STRIMZI_CLIENTS_CA_RENEWAL, "10");
 
-        Map labels = new HashMap<>(2);
+        Map<String, String> labels = new HashMap<>(2);
         labels.put("label1", "value1");
         labels.put("label2", "value2");
 
@@ -45,8 +48,10 @@ public class UserOperatorConfigTest {
         assertThat(config.getLabels(), is(expectedLabels));
         assertThat(config.getCaCertSecretName(), is(envVars.get(UserOperatorConfig.STRIMZI_CA_CERT_SECRET_NAME)));
         assertThat(config.getCaNamespace(), is(envVars.get(UserOperatorConfig.STRIMZI_CA_NAMESPACE)));
-        assertThat(config.getZookeperConnect(), is(envVars.get(UserOperatorConfig.STRIMZI_ZOOKEEPER_CONNECT)));
+        assertThat(config.getZookeeperConnect(), is(envVars.get(UserOperatorConfig.STRIMZI_ZOOKEEPER_CONNECT)));
         assertThat(config.getZookeeperSessionTimeoutMs(), is(Long.parseLong(envVars.get(UserOperatorConfig.STRIMZI_ZOOKEEPER_SESSION_TIMEOUT_MS))));
+        assertThat(config.getClientsCaValidityDays(), is(1000));
+        assertThat(config.getClientsCaRenewalDays(), is(10));
     }
 
     @Test
@@ -106,5 +111,16 @@ public class UserOperatorConfigTest {
         envVars.put(UserOperatorConfig.STRIMZI_LABELS, ",label1=");
 
         assertThrows(InvalidConfigurationException.class, () -> UserOperatorConfig.fromMap(envVars));
+    }
+
+    @Test
+    public void testFromMapCaValidityRenewalEnvVarMissingSetsDefault()  {
+        Map<String, String> envVars = new HashMap<>(UserOperatorConfigTest.envVars);
+        envVars.remove(UserOperatorConfig.STRIMZI_CLIENTS_CA_VALIDITY);
+        envVars.remove(UserOperatorConfig.STRIMZI_CLIENTS_CA_RENEWAL);
+
+        UserOperatorConfig config = UserOperatorConfig.fromMap(envVars);
+        assertThat(config.getClientsCaValidityDays(), is(CertificateAuthority.DEFAULT_CERTS_VALIDITY_DAYS));
+        assertThat(config.getClientsCaRenewalDays(), is(CertificateAuthority.DEFAULT_CERTS_RENEWAL_DAYS));
     }
 }